Separate out the key exchange enumeration into class KeyExchangeAlgorithm

+++ crypto/src/org/bouncycastle/crypto/tls/TlsECDHEKeyExchange.java
@@ -15,8 +15,7 @@ import org.bouncycastle.math.ec.ECPoint;
  */
 class TlsECDHEKeyExchange extends TlsECKeyExchange
 {
-    TlsECDHEKeyExchange(TlsClientContext context, CertificateVerifyer verifyer,
-        short keyExchange)
+    TlsECDHEKeyExchange(TlsClientContext context, CertificateVerifyer verifyer, int keyExchange)
     {
         super(context, verifyer, keyExchange);
     }
